enhancement to run scalatest test with ScalaRunner
currently the scalatest launcher uses standard java vm runner
to run or debug tests. It works fine with a java expression
evaluator but with the Scala IDE 4.1.0 there is available the
Scala Expression Evaluator which requires a Scala Debugger.
This enhancement adds a new scala launcher. From now it is
possible to choose appropriate launcher:
 - an old java one especially for java projects where tests are written
   with scalatest. Then debugging and evaluation can be done with a java
   evaluator
 - a new scala one for scala projects (can be used for java projects as
   well but is not documented).
You can stop in breakpoint during a test execution and evaluate
arbitrary expression.

This PR must be taken with scala-ide PR #955

Fix #1002447